--- src/main/kotlin/uk/gov/justice/digital/hmpps/approvedpremisesapi/api/ReferenceDataApiDelegate.kt
@@ -1,179 +1,0 @@
-package uk.gov.justice.digital.hmpps.approvedpremisesapi.api
-
-import org.springframework.http.HttpStatus
-import org.springframework.http.MediaType
-import org.springframework.http.ResponseEntity
-import org.springframework.web.context.request.NativeWebRequest
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.ApArea
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.CancellationReason
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.Characteristic
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.DepartureReason
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.DestinationProvider
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.LocalAuthorityArea
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.LostBedReason
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.MoveOnCategory
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.NonArrivalReason
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.ProbationDeliveryUnit
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.ProbationRegion
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.ReferralRejectionReason
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.ServiceName
-import java.util.Optional
-
-@jakarta.annotation.Generated(value = ["org.openapitools.codegen.languages.KotlinSpringServerCodegen"], comments = "Generator version: 7.13.0")
-interface ReferenceDataApiDelegate {
-
-  fun getRequest(): Optional<NativeWebRequest> = Optional.empty()
-
-  fun referenceDataApAreasGet(): ResponseEntity<List<ApArea>> {
-    getRequest().ifPresent { request ->
-      for (mediaType in MediaType.parseMediaTypes(request.getHeader("Accept"))) {
-        if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
-          ApiUtil.setExampleResponse(request, "application/json", "[ {  \"identifier\" : \"LON\",  \"name\" : \"Yorkshire & The Humber\",  \"id\" : \"cd1c2d43-0b0b-4438-b0e3-d4424e61fb6a\"}, {  \"identifier\" : \"LON\",  \"name\" : \"Yorkshire & The Humber\",  \"id\" : \"cd1c2d43-0b0b-4438-b0e3-d4424e61fb6a\"} ]")
-          break
-        }
-      }
-    }
-    return ResponseEntity(HttpStatus.NOT_IMPLEMENTED)
-  }
-
-  fun referenceDataCancellationReasonsGet(xServiceName: ServiceName?): ResponseEntity<List<CancellationReason>> {
-    getRequest().ifPresent { request ->
-      for (mediaType in MediaType.parseMediaTypes(request.getHeader("Accept"))) {
-        if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
-          ApiUtil.setExampleResponse(request, "application/json", "[ {  \"name\" : \"Recall\",  \"serviceScope\" : \"serviceScope\",  \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",  \"isActive\" : true}, {  \"name\" : \"Recall\",  \"serviceScope\" : \"serviceScope\",  \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",  \"isActive\" : true} ]")
-          break
-        }
-      }
-    }
-    return ResponseEntity(HttpStatus.NOT_IMPLEMENTED)
-  }
-
-  fun referenceDataCharacteristicsGet(
-    xServiceName: ServiceName?,
-    modelScope: String?,
-  ): ResponseEntity<List<Characteristic>> {
-    getRequest().ifPresent { request ->
-      for (mediaType in MediaType.parseMediaTypes(request.getHeader("Accept"))) {
-        if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
-          ApiUtil.setExampleResponse(request, "application/json", "[ {  \"propertyName\" : \"isCatered\",  \"name\" : \"Is this premises catered (rather than self-catered)?\",  \"serviceScope\" : \"approved-premises\",  \"id\" : \"952790c0-21d7-4fd6-a7e1-9018f08d8bb0\",  \"modelScope\" : \"premises\"}, {  \"propertyName\" : \"isCatered\",  \"name\" : \"Is this premises catered (rather than self-catered)?\",  \"serviceScope\" : \"approved-premises\",  \"id\" : \"952790c0-21d7-4fd6-a7e1-9018f08d8bb0\",  \"modelScope\" : \"premises\"} ]")
-          break
-        }
-      }
-    }
-    return ResponseEntity(HttpStatus.NOT_IMPLEMENTED)
-  }
-
-  fun referenceDataDepartureReasonsGet(
-    xServiceName: ServiceName?,
-    includeInactive: Boolean?,
-  ): ResponseEntity<List<DepartureReason>> {
-    getRequest().ifPresent { request ->
-      for (mediaType in MediaType.parseMediaTypes(request.getHeader("Accept"))) {
-        if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
-          ApiUtil.setExampleResponse(request, "application/json", "[ {  \"parentReasonId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",  \"name\" : \"Admitted to Hospital\",  \"serviceScope\" : \"serviceScope\",  \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",  \"isActive\" : true}, {  \"parentReasonId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",  \"name\" : \"Admitted to Hospital\",  \"serviceScope\" : \"serviceScope\",  \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",  \"isActive\" : true} ]")
-          break
-        }
-      }
-    }
-    return ResponseEntity(HttpStatus.NOT_IMPLEMENTED)
-  }
-
-  fun referenceDataDestinationProvidersGet(): ResponseEntity<List<DestinationProvider>> {
-    getRequest().ifPresent { request ->
-      for (mediaType in MediaType.parseMediaTypes(request.getHeader("Accept"))) {
-        if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
-          ApiUtil.setExampleResponse(request, "application/json", "[ {  \"name\" : \"Ext - North East Region\",  \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",  \"isActive\" : true}, {  \"name\" : \"Ext - North East Region\",  \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",  \"isActive\" : true} ]")
-          break
-        }
-      }
-    }
-    return ResponseEntity(HttpStatus.NOT_IMPLEMENTED)
-  }
-
-  fun referenceDataLocalAuthorityAreasGet(): ResponseEntity<List<LocalAuthorityArea>> {
-    getRequest().ifPresent { request ->
-      for (mediaType in MediaType.parseMediaTypes(request.getHeader("Accept"))) {
-        if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
-          ApiUtil.setExampleResponse(request, "application/json", "[ {  \"identifier\" : \"LEEDS\",  \"name\" : \"Leeds City Council\",  \"id\" : \"6abb5fa3-e93f-4445-887b-30d081688f44\"}, {  \"identifier\" : \"LEEDS\",  \"name\" : \"Leeds City Council\",  \"id\" : \"6abb5fa3-e93f-4445-887b-30d081688f44\"} ]")
-          break
-        }
-      }
-    }
-    return ResponseEntity(HttpStatus.NOT_IMPLEMENTED)
-  }
-
-  fun referenceDataLostBedReasonsGet(xServiceName: ServiceName?): ResponseEntity<List<LostBedReason>> {
-    getRequest().ifPresent { request ->
-      for (mediaType in MediaType.parseMediaTypes(request.getHeader("Accept"))) {
-        if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
-          ApiUtil.setExampleResponse(request, "application/json", "[ {  \"name\" : \"Double Room with Single Occupancy - Other (Non-FM)\",  \"serviceScope\" : \"serviceScope\",  \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",  \"isActive\" : true}, {  \"name\" : \"Double Room with Single Occupancy - Other (Non-FM)\",  \"serviceScope\" : \"serviceScope\",  \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",  \"isActive\" : true} ]")
-          break
-        }
-      }
-    }
-    return ResponseEntity(HttpStatus.NOT_IMPLEMENTED)
-  }
-
-  fun referenceDataMoveOnCategoriesGet(
-    xServiceName: ServiceName?,
-    includeInactive: Boolean?,
-  ): ResponseEntity<List<MoveOnCategory>> {
-    getRequest().ifPresent { request ->
-      for (mediaType in MediaType.parseMediaTypes(request.getHeader("Accept"))) {
-        if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
-          ApiUtil.setExampleResponse(request, "application/json", "[ {  \"name\" : \"Housing Association - Rented\",  \"serviceScope\" : \"serviceScope\",  \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",  \"isActive\" : true}, {  \"name\" : \"Housing Association - Rented\",  \"serviceScope\" : \"serviceScope\",  \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",  \"isActive\" : true} ]")
-          break
-        }
-      }
-    }
-    return ResponseEntity(HttpStatus.NOT_IMPLEMENTED)
-  }
-
-  fun referenceDataNonArrivalReasonsGet(): ResponseEntity<List<NonArrivalReason>> {
-    getRequest().ifPresent { request ->
-      for (mediaType in MediaType.parseMediaTypes(request.getHeader("Accept"))) {
-        if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
-          ApiUtil.setExampleResponse(request, "application/json", "[ {  \"name\" : \"Recall\",  \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",  \"isActive\" : true}, {  \"name\" : \"Recall\",  \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",  \"isActive\" : true} ]")
-          break
-        }
-      }
-    }
-    return ResponseEntity(HttpStatus.NOT_IMPLEMENTED)
-  }
-
-  fun referenceDataProbationDeliveryUnitsGet(probationRegionId: java.util.UUID?): ResponseEntity<List<ProbationDeliveryUnit>> {
-    getRequest().ifPresent { request ->
-      for (mediaType in MediaType.parseMediaTypes(request.getHeader("Accept"))) {
-        if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
-          ApiUtil.setExampleResponse(request, "application/json", "[ {  \"name\" : \"name\",  \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\"}, {  \"name\" : \"name\",  \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\"} ]")
-          break
-        }
-      }
-    }
-    return ResponseEntity(HttpStatus.NOT_IMPLEMENTED)
-  }
-
-  fun referenceDataProbationRegionsGet(): ResponseEntity<List<ProbationRegion>> {
-    getRequest().ifPresent { request ->
-      for (mediaType in MediaType.parseMediaTypes(request.getHeader("Accept"))) {
-        if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
-          ApiUtil.setExampleResponse(request, "application/json", "[ {  \"name\" : \"NPS North East Central Referrals\",  \"id\" : \"952790c0-21d7-4fd6-a7e1-9018f08d8bb0\"}, {  \"name\" : \"NPS North East Central Referrals\",  \"id\" : \"952790c0-21d7-4fd6-a7e1-9018f08d8bb0\"} ]")
-          break
-        }
-      }
-    }
-    return ResponseEntity(HttpStatus.NOT_IMPLEMENTED)
-  }
-
-  fun referenceDataReferralRejectionReasonsGet(xServiceName: ServiceName?): ResponseEntity<List<ReferralRejectionReason>> {
-    getRequest().ifPresent { request ->
-      for (mediaType in MediaType.parseMediaTypes(request.getHeader("Accept"))) {
-        if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
-          ApiUtil.setExampleResponse(request, "application/json", "[ {  \"name\" : \"There was not enough time to place them\",  \"serviceScope\" : \"serviceScope\",  \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",  \"isActive\" : true}, {  \"name\" : \"There was not enough time to place them\",  \"serviceScope\" : \"serviceScope\",  \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",  \"isActive\" : true} ]")
-          break
-        }
-      }
-    }
-    return ResponseEntity(HttpStatus.NOT_IMPLEMENTED)
-  }
-}
